Extract shared LabelMode type to internal/viz package
Move the LabelMode type and its constants (LabelAll, LabelFoldersOnly,
LabelLaps, LabelNone) to a new internal/viz package as the single source
of truth. Each layout package (radialtree, bubbletree, spiral) re-exports
the type via a type alias and the constants it uses, preserving full
backward compatibility for all consumers.

Closes #154
